+++ crates/grpc/src/handlers/mod.rs
@@ -4,7 +4,5 @@
 //! The handlers delegate to the underlying `StarknetApi` implementation for business logic.
 
 mod starknet;
-mod trace;
-mod write;
 
 pub use self::starknet::StarknetService;
